refactor(honeycrisp): extract self-contained FolderMenuItem, remove dead folderName branch
Extract FolderMenuItem.svelte from Sidebar per the self-contained component
pattern — each folder row now owns its own rename editing state and delete
confirmation dialog instead of sharing a single dialog managed by the parent.
Sidebar drops from 212 to 93 lines. Also remove the dead isRecentlyDeletedView
branch from viewState.folderName since the parent now passes the title directly.
